Reimprovise

+++ Sections/ManageItems/ViewPerItem_Form.cs
@@ -16,18 +16,28 @@ namespace MyWinFormsApp.Sections.ManageItems
         ManageItems_form parent;
         Sqlsupportlocal sql = new Sqlsupportlocal(".\\SQLEXPRESS", "TruckEstimationSystem", null, null);
         int id;
-        public ViewPerItem_Form(string clientname, string itemname, decimal length, decimal width, decimal height, string fluteCode, int id, ManageItems_form parent)
+        public string itemname = "";
+        public string clientname = "";
+        public string flutecode = "";
+        public string category = "";
+        public ViewPerItem_Form(string clientname, string itemname, decimal length, decimal width, decimal height, string fluteCode, int id, string category, string fcc, ManageItems_form parent, bool giveWarning = false)
         {
             InitializeComponent();
             this.parent = parent;
             this.id = id;
-            client_label.Text = $"Client: {clientname}";
+            client_label.Text = $"{clientname}";
             itemname_label.Text = itemname;
-            flute_label.Text = $"FluteType: {fluteCode}";
-            content_label.Text = $"Length(mm):     {length}\n" +
-                                 $"Width(mm) :     {width}\n" +
-                                 $"Height(mm):     {height}\n" +
-                                 $"Dimension (mm): {height * width * width}";
+            flute_label.Text = $"FluteType: ({fluteCode.ToUpper()})";
+            content_label.Text = $"Length(mm):     {Math.Round(length,2)}\n" +
+                                 $"Width(mm) :     {Math.Round(width,2)}\n" +
+                                 $"Height(mm):     {Math.Round(height,2)}\n" +
+                                 $"Dimension (mm): {Math.Round(height * width * height, 0)}";
+            this.fcc_label.Text = $"F.C. CONTROL No.:\n    {fcc}";
+            this.category = category;
+            this.itemname = itemname;
+            this.clientname = clientname;
+            this.flutecode = fluteCode;
+            if (giveWarning) this.BackColor = Color.MistyRose;
         }
 
         private void tableLayoutPanel1_Paint(object sender, PaintEventArgs e)
@@ -37,8 +47,21 @@ namespace MyWinFormsApp.Sections.ManageItems
 
         private void delete_btn_Click(object sender, EventArgs e)
         {
-            DeleteMyValue();
-            parent.UpdateVisual();
+            DialogResult result = MessageBox.Show(
+"Once this data is deleted, it can no longer be retrieve. Are you sure?",
+"Warning",
+MessageBoxButtons.YesNo,
+MessageBoxIcon.Exclamation
+);
+            if (result == DialogResult.Yes)
+            {
+                sql.commitReport($"A data Item '{itemname}' was deleted");
+                DeleteMyValue();
+                parent.UpdateVisual();
+            }
+            else if (result == DialogResult.No)
+            {
+            }
         }
 
         private void edit_btn_Click(object sender, EventArgs e)
